+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/data/ChannelAttendeesCount.kt
@@ -5,18 +5,6 @@ package jp.shiguredo.sora.sdk.channel.data
  */
 data class ChannelAttendeesCount(
     /**
-     * 配信者数.
-     */
-    @Deprecated("numberOfUpstreams は 2021 年 6 月リリース予定の Sora にて廃止されます。")
-    val numberOfUpstreams: Int,
-
-    /**
-     * 視聴者数.
-     */
-    @Deprecated("numberOfDownstreams は 2021 年 6 月リリース予定の Sora にて廃止されます。")
-    val numberOfDownstreams: Int,
-
-    /**
      * sendrecv の接続数.
      */
     val numberOfSendrecvConnections: Int,